pyproject.toml
[tool.poetry]
name = "soccon"
version = "0.1.0"
description = ""
authors = ["Nelson Barnett"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
xlsxwriter = "^3.2.0"
matplotlib = "^3.10.0"
mano = "^0.5.2"
beiwe-forest = {git = "https://github.com/onnela-lab/forest.git"}
orjson = "^3.10.15"

[tool.poetry.scripts]
soccon_analysis = "soccon.main:main"
data_check = "soccon.quality_check:download_funcs_cli"
run_quality_check = "soccon.quality_check:quality_check_cli"
make_key = "soccon.make_key:main"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
